authentication/drf/api/requirements.txt
Django==2.2.7
djangorestframework==3.10.3
djangorestframework_simplejwt==4.3.0
django-cors-headers==3.1.1
djoser==2.0.3
gunicorn==19.9.0
mysqlclient==1.4.5

